+++ Clubber.Domain/Helpers/DatabaseHelper.cs
@@ -23,7 +23,7 @@ public class DatabaseHelper : IDatabaseHelper
 	{
 		try
 		{
-			DdUser newDdUser = new(discordId, (int)lbId);
+			DdUser newDdUser = new(discordId, lbId);
 
 			await _dbContext.AddAsync(newDdUser);
 			await _dbContext.SaveChangesAsync();
